Revert "Revert "MORE CHAOS""
This reverts commit d70d672053f76b3364bc36304bba53d912076b4b.

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: htrindad <htrindad@student.42lisboa.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/06 21:03:23 by htrindad          #+#    #+#             */
-/*   Updated: 2024/05/06 21:03:42 by htrindad         ###   ########.fr       */
+/*   Updated: 2024/05/10 16:34:52 by htrindad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i])
